#82 refactor: create visualization module, collapse engine + types
- Create src/lib/modules/visualization/ absorbing all engine files + tree_visualization.ts
- Add computeVisualization pipeline (mapIssueToStateDimensions + computeTreeVisualization)
- Keep computeForestLayout as separate entry point
- Merge all 4 test files into visualization.test.ts (120 tests)
- Delete src/lib/engine/ directory and types/tree_visualization.ts
